refactor(web): F3 后台任务桶下沉 useTaskBuckets hook
桶状态（taskMapRef/tasks/tasksOpen）与辅助群（pubTasks/taskBucket/appendTaskMsg/
appendSidechain/markTerminal/settleBucketFromResult/maybeFetchCodexTranscript/
hydrateTasks）+ E4 TTL 驱逐逐字迁入 hooks/useTaskBuckets.ts；api 为每渲染重建的
普通对象（内部全走 ref/稳定 setState/isCodex 会话内不变，过期闭包语义等价）。
resetFromHistory 收容 applyHistory 的水合段；task_started 的桌面端自动开栏随迁。
Chat.tsx 1269→1158 行。零行为改动。
